Illustrate the manual with live components rather than screenshots
Extract the OBS Builder's preview into OBSFieldPreview, and use it in the
manual's Fields chapter to show what score, names, unseen tiles, blanks,
spread and the scrolling last play actually look like. It is the same
component the builder previews with and the same sample data, so the manual
cannot show something the product no longer does.

Woogles UI in the manual is embedded live for the same reason -- the
'streamed by 2' tag in the director chapter is a real antd Tag. An embed
cannot go stale, follows light and dark mode for free, and adds no binary
to review. This manual has already needed two corrections for claims that
drifted from the code; screenshots have that failure mode with nothing to
catch it.

That leaves screenshots for the one thing we cannot render: OBS itself.
Three DocImage slots mark where they go. DocImage renders a labelled
placeholder when the file is missing, so the manual reads sensibly until
the images land and a deleted file is obvious rather than silently blank.
