+++ liwords-ui/src/docs/prose.tsx
@@ -22,3 +22,37 @@ export const H: React.FC<{ children: React.ReactNode }> = ({ children }) => (
     {children}
   </Typography.Title>
 );
+
+/**
+ * A screenshot, for the few things the app cannot render itself — chiefly OBS's
+ * own interface.
+ *
+ * Prefer embedding the real component over a screenshot wherever the subject is
+ * Woogles UI: an embed cannot go stale, follows light and dark mode, and needs
+ * no binary in the repo. Reach for this only when the subject is somebody
+ * else's application.
+ *
+ * Renders a labelled placeholder when the file is missing, so a manual can be
+ * written before its screenshots are taken and a deleted file is obvious rather
+ * than silently blank.
+ */
+export const DocImage: React.FC<{
+  src: string;
+  alt: string;
+  caption?: string;
+}> = ({ src, alt, caption }) => {
+  const [failed, setFailed] = React.useState(false);
+
+  return (
+    <figure className="doc-figure">
+      {failed ? (
+        <div className="doc-figure-missing">
+          Screenshot not available: <code>{src}</code>
+        </div>
+      ) : (
+        <img src={src} alt={alt} onError={() => setFailed(true)} />
+      )}
+      {caption && <figcaption>{caption}</figcaption>}
+    </figure>
+  );
+};
